requirements.txt
streamlit
langchain-core
langchain-groq
google-api-python-client
google-auth-oauthlib
python-dotenv
requests